chore : inputFileds 상수 파일로 분리
