+++ src/pages/SignUp/SignUpKakao.jsx
@@ -5,7 +5,7 @@ import GlobalStyle from '@/styles/global';
 import { useNavigate, useLocation } from 'react-router-dom';
 import { KakaoSignUp } from '@/api/SignUp';
 import { useMutation } from '@tanstack/react-query';
-import { createInputFields } from '@/constant/SignupFields';
+import { createInputFields } from '@/utils/SignupFields';
 import { onValidation } from '@/utils/validation';
 import { Container, Content, Text, SocialBox } from '@/styles/SignupStyles';
 import { useToast } from '@/context/ToastContext';
